% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvfunctions.R
\name{pvfunctions}
\alias{pvfunctions}
\alias{pv}
\alias{fv}
\alias{npv}
\alias{irr}
\alias{pmt}
\alias{ipmt}
\alias{ppmt}
\title{pvfunctions}
\usage{
pv(rate,nper,pymt,fvamt)
npv(rate, values)
irr(values, lower = -0.5, upper = 1.5, tolerance = .Machine$double.eps^0.25)
pmt(rate,nper,pvamt, fvamt)
ipmt(rate, period, nper, pvamt, fvamt)
ppmt(rate, period,nper, pvamt, fvamt)
}
\arguments{
\item{rate}{effective per-period interest rate}

\item{nper}{number of periods}

\item{pymt}{periodic payment (e.g. annuity)}

\item{fvamt}{future value amount not accounted for by periodic
payment.}

\item{pvamt}{present amount (e.g. amt to be annuitized); in case of \code{FV}, specifies current amount paid on annuity}

\item{period}{specific period for which calculation is to be
performed (e.g. interest component of annuity)}

\item{values}{vector of payments}

\item{lower}{lower bound of search for irr}

\item{upper}{upper bound of search for irr}

\item{tolerance}{convergence tolerance for irr. Set to \code{uniroot} default}
}
\description{
Experimental implemention of basic present value spreadsheet
functions. The functions \code{pv}, \code{fv}, \code{pmt}, \code{ipmt}, \code{ppmt}, deal
with annuities (reqularly repeating payments). The functions
\code{irr} and \code{npv} deal with unequal cash flows, with \code{npv}
assuming the first payment occurs in one period. Specifically,
\itemize{
\item \code{pv} and \code{fv} compute the present and future value of an
annuity. The optional argument (which defaults to 0) permits an
ongoing annuity to have a current balance (with the \code{fv}
function) or for there to be a future balance (with the \code{pv}
function)
\item \code{pmt}, \code{ipmt}, and \code{ppmt} compute the annuity
payment, and for any given payment the breakdown between
interest and principal
\item \code{npv} and \code{irr} accept a vector as input to handle cash flows that are not annuities.
\item These functions should adhere to the sign conventions a spreadsheet user would expect.
}

\code{pv(rate,nper,pymt,0 )}
\code{pv(rate,nper,pymt,fvamt)}
\code{fv(rate,nper,pymt,0 )}
\code{fv(rate,nper,pymt,pvamt)}
\code{npv(rate,values )}
\code{irr(values2, lower = -0.5, upper = 1.5, tolerance = .Machine$double.eps^0.25)}
\code{pmt(rate,nper,pvamt, fvamt=0)}
\code{ipmt(rate, period, nper, pvamt)}
\code{ppmt(rate, period,nper, pvamt)}
}
\details{
These functions (mostly) mimic the behavior of the
standard basic financial function in Excel, Google Sheets, and
Libreoffice. With the exception of \code{irr}, they are
vectorized.
}
